capstone_project/requirements.txt
streamlit
python-dotenv
PyPDF2
python-docx
requests
streamlit-lottie
fpdf
google-genai
